+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `foreach` support in `SkipStepAdamW`.
 - Added `budget` mode for activation checkpointing configuration.
 - Added `io.remove_file()` function.
+- Added ABF, PI, and YaRN rope scaling strategies.
 
 ### Changed